// client/components/saved/__generated__/SavedJobsListInternalFragment.graphql.ts
export type SavedJobsListInternalFragment$data = {
  readonly savedJobs: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"JobFragment">;
      };
    }>;
    readonly pageInfo: {
      readonly hasNextPage: boolean;
    };
  };
  readonly " $fragmentType": "SavedJobsListInternalFragment";
};
export type SavedJobsListInternalFragment$key = {
  readonly " $data"?: SavedJobsListInternalFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"SavedJobsListInternalFragment">;
};